# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface FetchAppJsonBody {
  /** GitHub repository URL (https://github.com/owner/repo) */
  repoUrl: string;
}

export interface AppJsonEnvVar {
  description?: string;
  value?: string;
  required?: boolean;
  generator?: string;
}

export type AppJsonDataEnv = { [key: string]: AppJsonEnvVar };

export type AppJsonDataScripts = { [key: string]: string };

export interface AppJsonData {
  name: string;
  description?: string;
  keywords?: string[];
  env?: AppJsonDataEnv;
  scripts?: AppJsonDataScripts;
}

export interface App {
  id: number;
  name: string;
  description?: string;
  repoUrl: string;
  repoOwner: string;
  repoName: string;
  appJson: AppJsonData;
  deploymentCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAppBody {
  /** GitHub repository URL */
  repoUrl: string;
}

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  available: "available",
  occupied: "occupied",
  maintenance: "maintenance",
} as const;

export type DeploymentStatus =
  (typeof DeploymentStatus)[keyof typeof DeploymentStatus];

export const DeploymentStatus = {
  pending: "pending",
  running: "running",
  stopped: "stopped",
  failed: "failed",
  restarting: "restarting",
} as const;

export type DeploymentEnvConfig = { [key: string]: string };

export interface Deployment {
  id: number;
  appId: number;
  serverId: number;
  status: DeploymentStatus;
  envConfig: DeploymentEnvConfig;
  deployedBy?: string | null;
  app?: App | null;
  server?: Server | null;
  createdAt: string;
  updatedAt: string;
}

export interface Server {
  id: number;
  slotNumber: number;
  label: string;
  status: ServerStatus;
  region: string;
  currentDeploymentId?: number | null;
  currentDeployment?: Deployment | null;
  createdAt: string;
}

export interface ServerStats {
  total: number;
  available: number;
  occupied: number;
  maintenance: number;
}

/**
 * Environment variable values from the app.json form
 */
export type CreateDeploymentBodyEnvConfig = { [key: string]: string };

export interface CreateDeploymentBody {
  /** Optional app ID. If omitted, the platform's configured bot is used. */
  appId?: number | null;
  /** Optional specific server slot ID. If omitted, auto-assigns a free slot. */
  serverId?: number | null;
  /** Environment variable values from the app.json form */
  envConfig: CreateDeploymentBodyEnvConfig;
  /** Name or handle of the person deploying */
  deployedBy?: string | null;
}

export interface PlatformConfig {
  botRepoUrl: string;
  botRepoOwner: string;
  botRepoName: string;
  botName: string;
  botDescription?: string | null;
  botLogo?: string | null;
  botAppJson: AppJsonData;
  slotCount: number;
  totalSlots: number;
  occupiedSlots: number;
  availableSlots: number;
  updatedAt: string;
  adminPasswordIsDefault: boolean;
}

export interface UpdatePlatformConfigBody {
  /** New GitHub repo URL for the bot */
  botRepoUrl?: string;
  /**
   * @minimum 1
   * @maximum 500
   */
  slotCount?: number;
}

export interface ActiveSessions {
  count: number;
  fingerprints: string[];
}

export interface DeploymentSummary {
  running: number;
  stopped: number;
  pending: number;
  failed: number;
  total: number;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
} as const;

export interface LogEntry {
  timestamp: string;
  level: LogEntryLevel;
  message: string;
}

export interface DeploymentLogs {
  deploymentId: number;
  logs: LogEntry[];
}

export type PlatformLoginBody = {
  password: string;
};